# contracts/openapi/agent-api.yaml
openapi: 3.1.0
info:
  title: SGC AI Agent API
  version: 0.1.0
  description: Smart Garage Customer AI Agent — plug-and-play API

servers:
  - url: http://localhost:8000
    description: Local development

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    CustomerContext:
      type: object
      properties:
        customer_id: { type: string }
        vehicle_id: { type: string }
        vehicle_make: { type: string }
        vehicle_model: { type: string }
        vehicle_variant: { type: string }
        fuel_type: { type: string }
        registration_no: { type: string }
        mileage_km: { type: integer }
        vehicle_age_months: { type: integer }
        locale: { type: string, default: en-IN }

    LocationContext:
      type: object
      properties:
        latitude: { type: number }
        longitude: { type: number }
        area: { type: string }

    ContextEnvelope:
      type: object
      properties:
        customer: { $ref: '#/components/schemas/CustomerContext' }
        location: { $ref: '#/components/schemas/LocationContext' }
        garage_id: { type: string }
        metadata: { type: object, additionalProperties: { type: string } }

    CreateSessionRequest:
      type: object
      properties:
        customer_id: { type: string }
        vehicle_id: { type: string }
        context: { $ref: '#/components/schemas/ContextEnvelope' }

    CreateSessionResponse:
      type: object
      required: [session_id]
      properties:
        session_id: { type: string }

    ChatRequest:
      type: object
      required: [session_id, message]
      properties:
        session_id: { type: string }
        message: { type: string }
        context: { $ref: '#/components/schemas/ContextEnvelope' }

    ChatResponse:
      type: object
      required: [session_id, response]
      properties:
        session_id: { type: string }
        response: { type: string }
        intent: { type: string }
        source_refs: { type: array, items: { type: object } }
        requires_human_review: { type: boolean }
        model_used: { type: string }

    ToolInvokeRequest:
      type: object
      required: [tool_name]
      properties:
        tool_name: { type: string }
        parameters: { type: object }
        session_id: { type: string }

    ToolInvokeResponse:
      type: object
      properties:
        tool_name: { type: string }
        success: { type: boolean }
        data: {}
        error: { type: string }

paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: OK

  /v1/models:
    get:
      summary: List available LLM models
      responses:
        '200':
          description: Model list

  /v1/sessions:
    post:
      summary: Create conversation session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateSessionResponse' }

  /v1/chat:
    post:
      summary: Send chat message
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatRequest' }
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatResponse' }

  /v1/tools/invoke:
    post:
      summary: Directly invoke an agent tool
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ToolInvokeRequest' }
      responses:
        '200':
          description: Tool result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ToolInvokeResponse' }